refactor: make tooltip as internal component
