--- src/client/theme-default/slots/Tooltip/index.less
@@ -1,46 +1,0 @@
-.dumi-theme-default-tooltip {
-  position: fixed;
-  min-width: 30px;
-  margin-bottom: 8px;
-  box-sizing: border-box;
-
-  &-hidden {
-    display: none;
-  }
-
-  &-arrow {
-    border: 4px solid transparent;
-    opacity: 0.7;
-  }
-
-  &-placement-top {
-    .dumi-theme-default-tooltip-arrow {
-      border-top-color: #000;
-      transform: translate(-50%, 100%);
-    }
-  }
-
-  &-bottom {
-    .dumi-theme-default-tooltip-arrow {
-      border-bottom-color: #000;
-      transform: translateY(-50%, -100%);
-    }
-  }
-
-  &-content {
-    padding: 5px 8px;
-    color: #fff;
-    font-size: 13px;
-    line-height: 1.1;
-    white-space: nowrap;
-    background-color: #000;
-    opacity: 0.7;
-    border-radius: 2px;
-    box-shadow: 0 4px 10px rgba(0, 0, 0, 10%);
-    box-sizing: border-box;
-
-    @{dark-selector} & {
-      background-color: #222;
-    }
-  }
-}
